TRI-FACTOR STOCK EXCHANGE
For every business weather it is small scale business or medium scale
business or large scale business needs raw material. Coming to the point
in the world 77% business are small scale business. Here we are going to
develop a mobile app for small scale businesses from a place needs to
get somethings like raw materials. So, four to five stores get together
to get things which they need cheap from a manufacture or wholesaler